+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## ScottPlot 5.0.27
 _Not yet on NuGet..._
-* SignalXY: Corrected floating point arithmetic error that caused points to be dropped (#3665) @mjazd
+* Signal: Corrected floating point error that caused points to be dropped in rare conditions (#3665) @mjazd
 
 ## ScottPlot 5.0.26
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-04-14_